refactor(client): split studio, join, and health pages into sections
Extract use-studio-page and use-join-room-page hooks with focused panels so
chat and join routes stay maintainable under react-doctor size limits.
